requirements.txt
matplotlib==3.4.1
numpy >=1.8
pandas<1.2.3
xlrd<2.0
streamlit==0.79.0
plotly==4.14.3
Pillow==8.1.2
